requirements.txt
mysql-connector-python
pyarrow